Remove dialog when accepting safety number changes
