+++ src/io/forsta/securesms/ConversationItem.java
@@ -515,111 +515,111 @@ public class ConversationItem extends LinearLayout
 //    new ConfirmIdentityDialog(context, masterSecret, messageRecord, mismatches.get(0)).show();
   }
 
-  private class AcceptIdentityMismatch extends AsyncTask<Void, Void, Void> {
-    private final Context context;
-    private final MasterSecret        masterSecret;
-    private final MessageRecord       messageRecord;
-    private final IdentityKeyMismatch mismatch;
-
-    public AcceptIdentityMismatch(Context context, MasterSecret masterSecret, MessageRecord messageRecord, IdentityKeyMismatch mismatch) {
-      this.context = context;
-      this.masterSecret  = masterSecret;
-      this.messageRecord = messageRecord;
-      this.mismatch      = mismatch;
-    }
-
-    @Override
-    protected Void doInBackground(Void... voids) {
-      IdentityDatabase identityDatabase = DatabaseFactory.getIdentityDatabase(context);
-
-      identityDatabase.saveIdentity(mismatch.getRecipientId(),
-          mismatch.getIdentityKey());
-
-      processMessageRecord(messageRecord);
-      processPendingMessageRecords(messageRecord.getThreadId(), mismatch);
-
-      ApplicationContext.getInstance(context)
-          .getJobManager()
-          .add(new IdentityUpdateJob(context, mismatch.getRecipientId()));
-
-      return null;
-    }
-
-    private void processMessageRecord(MessageRecord messageRecord) {
-      if (messageRecord.isOutgoing()) processOutgoingMessageRecord(messageRecord);
-      else                            processIncomingMessageRecord(messageRecord);
-    }
-
-    private void processPendingMessageRecords(long threadId, IdentityKeyMismatch mismatch) {
-      MmsSmsDatabase mmsSmsDatabase = DatabaseFactory.getMmsSmsDatabase(getContext());
-      Cursor cursor         = mmsSmsDatabase.getIdentityConflictMessagesForThread(threadId);
-      MmsSmsDatabase.Reader reader         = mmsSmsDatabase.readerFor(cursor, masterSecret);
-      MessageRecord         record;
-
-      try {
-        while ((record = reader.getNext()) != null) {
-          for (IdentityKeyMismatch recordMismatch : record.getIdentityKeyMismatches()) {
-            if (mismatch.equals(recordMismatch)) {
-              processMessageRecord(record);
-            }
-          }
-        }
-      } finally {
-        if (reader != null)
-          reader.close();
-      }
-    }
-
-    private void processOutgoingMessageRecord(MessageRecord messageRecord) {
-      SmsDatabase smsDatabase        = DatabaseFactory.getSmsDatabase(getContext());
-      MmsDatabase        mmsDatabase        = DatabaseFactory.getMmsDatabase(getContext());
-      MmsAddressDatabase mmsAddressDatabase = DatabaseFactory.getMmsAddressDatabase(getContext());
-
-      if (messageRecord.isMms()) {
-        mmsDatabase.removeMismatchedIdentity(messageRecord.getId(),
-            mismatch.getRecipientId(),
-            mismatch.getIdentityKey());
-
-        Recipients recipients = mmsAddressDatabase.getRecipientsForId(messageRecord.getId());
-
-        if (recipients.isGroupRecipient()) MessageSender.resendGroupMessage(getContext(), masterSecret, messageRecord, mismatch.getRecipientId());
-        else                               MessageSender.resend(getContext(), masterSecret, messageRecord);
-      } else {
-        smsDatabase.removeMismatchedIdentity(messageRecord.getId(),
-            mismatch.getRecipientId(),
-            mismatch.getIdentityKey());
-
-        MessageSender.resend(getContext(), masterSecret, messageRecord);
-      }
-    }
-
-    private void processIncomingMessageRecord(MessageRecord messageRecord) {
-      try {
-        PushDatabase pushDatabase = DatabaseFactory.getPushDatabase(getContext());
-        SmsDatabase  smsDatabase  = DatabaseFactory.getSmsDatabase(getContext());
-
-        smsDatabase.removeMismatchedIdentity(messageRecord.getId(),
-            mismatch.getRecipientId(),
-            mismatch.getIdentityKey());
-
-        SignalServiceEnvelope envelope = new SignalServiceEnvelope(SignalServiceProtos.Envelope.Type.PREKEY_BUNDLE_VALUE,
-            messageRecord.getIndividualRecipient().getNumber(),
-            messageRecord.getRecipientDeviceId(), "",
-            messageRecord.getDateSent(),
-            Base64.decode(messageRecord.getBody().getBody()),
-            null);
-
-        long pushId = pushDatabase.insert(envelope);
-
-        ApplicationContext.getInstance(getContext())
-            .getJobManager()
-            .add(new PushDecryptJob(getContext(), pushId, messageRecord.getId(),
-                messageRecord.getIndividualRecipient().getNumber()));
-      } catch (IOException e) {
-        throw new AssertionError(e);
-      }
-    }
-  }
+//  private class AcceptIdentityMismatch extends AsyncTask<Void, Void, Void> {
+//    private final Context context;
+//    private final MasterSecret        masterSecret;
+//    private final MessageRecord       messageRecord;
+//    private final IdentityKeyMismatch mismatch;
+//
+//    public AcceptIdentityMismatch(Context context, MasterSecret masterSecret, MessageRecord messageRecord, IdentityKeyMismatch mismatch) {
+//      this.context = context;
+//      this.masterSecret  = masterSecret;
+//      this.messageRecord = messageRecord;
+//      this.mismatch      = mismatch;
+//    }
+//
+//    @Override
+//    protected Void doInBackground(Void... voids) {
+//      IdentityDatabase identityDatabase = DatabaseFactory.getIdentityDatabase(context);
+//
+//      identityDatabase.saveIdentity(mismatch.getRecipientId(),
+//          mismatch.getIdentityKey());
+//
+//      processMessageRecord(messageRecord);
+//      processPendingMessageRecords(messageRecord.getThreadId(), mismatch);
+//
+//      ApplicationContext.getInstance(context)
+//          .getJobManager()
+//          .add(new IdentityUpdateJob(context, mismatch.getRecipientId()));
+//
+//      return null;
+//    }
+//
+//    private void processMessageRecord(MessageRecord messageRecord) {
+//      if (messageRecord.isOutgoing()) processOutgoingMessageRecord(messageRecord);
+//      else                            processIncomingMessageRecord(messageRecord);
+//    }
+//
+//    private void processPendingMessageRecords(long threadId, IdentityKeyMismatch mismatch) {
+//      MmsSmsDatabase mmsSmsDatabase = DatabaseFactory.getMmsSmsDatabase(getContext());
+//      Cursor cursor         = mmsSmsDatabase.getIdentityConflictMessagesForThread(threadId);
+//      MmsSmsDatabase.Reader reader         = mmsSmsDatabase.readerFor(cursor, masterSecret);
+//      MessageRecord         record;
+//
+//      try {
+//        while ((record = reader.getNext()) != null) {
+//          for (IdentityKeyMismatch recordMismatch : record.getIdentityKeyMismatches()) {
+//            if (mismatch.equals(recordMismatch)) {
+//              processMessageRecord(record);
+//            }
+//          }
+//        }
+//      } finally {
+//        if (reader != null)
+//          reader.close();
+//      }
+//    }
+//
+//    private void processOutgoingMessageRecord(MessageRecord messageRecord) {
+//      SmsDatabase smsDatabase        = DatabaseFactory.getSmsDatabase(getContext());
+//      MmsDatabase        mmsDatabase        = DatabaseFactory.getMmsDatabase(getContext());
+//      MmsAddressDatabase mmsAddressDatabase = DatabaseFactory.getMmsAddressDatabase(getContext());
+//
+//      if (messageRecord.isMms()) {
+//        mmsDatabase.removeMismatchedIdentity(messageRecord.getId(),
+//            mismatch.getRecipientId(),
+//            mismatch.getIdentityKey());
+//
+//        Recipients recipients = mmsAddressDatabase.getRecipientsForId(messageRecord.getId());
+//
+//        if (recipients.isGroupRecipient()) MessageSender.resendGroupMessage(getContext(), masterSecret, messageRecord, mismatch.getRecipientId());
+//        else                               MessageSender.resend(getContext(), masterSecret, messageRecord);
+//      } else {
+//        smsDatabase.removeMismatchedIdentity(messageRecord.getId(),
+//            mismatch.getRecipientId(),
+//            mismatch.getIdentityKey());
+//
+//        MessageSender.resend(getContext(), masterSecret, messageRecord);
+//      }
+//    }
+//
+//    private void processIncomingMessageRecord(MessageRecord messageRecord) {
+//      try {
+//        PushDatabase pushDatabase = DatabaseFactory.getPushDatabase(getContext());
+//        SmsDatabase  smsDatabase  = DatabaseFactory.getSmsDatabase(getContext());
+//
+//        smsDatabase.removeMismatchedIdentity(messageRecord.getId(),
+//            mismatch.getRecipientId(),
+//            mismatch.getIdentityKey());
+//
+//        SignalServiceEnvelope envelope = new SignalServiceEnvelope(SignalServiceProtos.Envelope.Type.PREKEY_BUNDLE_VALUE,
+//            messageRecord.getIndividualRecipient().getNumber(),
+//            messageRecord.getRecipientDeviceId(), "",
+//            messageRecord.getDateSent(),
+//            Base64.decode(messageRecord.getBody().getBody()),
+//            null);
+//
+//        long pushId = pushDatabase.insert(envelope);
+//
+//        ApplicationContext.getInstance(getContext())
+//            .getJobManager()
+//            .add(new PushDecryptJob(getContext(), pushId, messageRecord.getId(),
+//                messageRecord.getIndividualRecipient().getNumber()));
+//      } catch (IOException e) {
+//        throw new AssertionError(e);
+//      }
+//    }
+//  }
 
   @Override
   public void onModified(final Recipient recipient) {
